hello-world-1/requirements.txt
functions-framework==3.5.0
requests
PyJWT